Use distroless Docker image
Refactor to support Google's distroless OpenJDK image. (See
https://github.com/GoogleContainerTools/distroless for details.)

Now we can do a `docker run` using the `--read-only` flag to make sure
the root image is read-only and only expose volumes we want
writable (e.g. /data).

Sadly, /tmp needs to be writable as apparently JNA uses it, so /tmp is
now listed as a volume.