Prototype of a local filesystem implementation
Firstly, refactor CloudUri to RemoteUri...no need to be "cloud" specific
per se.

Stub out a basic idea for an interface (AssetProvider) to abstract out
different backends. Adds in a very simple "registration" concept that
could be done either at runtime via reflection or in static blocks in
the provider implementations.

Add a simple naive test for now that needs to be refactored: simply try
to load a given private key using a LocalFSProvider RemoteUri.

Needs better tests, specifically of the hot-wired SslPolicyLoader.