Update SslPolicyLoaderTest
This commit updates the test to serve as an integration test making sure
the AssetUri setting type works.

Unfortunately, since the guts of the Config framework are foreign to me,
I found I had to override yet-another-class to get this to work. As such
the ContainerSslPolicyConfig was renamed to conflict with the built-in
SslPolicyConfig.