Fix Docker issues, refactoring Go module
As I move towards just having Gradle drive most of this, I've tucked the
Go code into a bit of an overkill module layout (cmd/gojava). However,
I've updated the Makefile to pull the project version from Gradle and
pass it to docker-build.

This also updates the Go building in the Dockerfiles for the new layout.
Plus incorporates the changes in using the dist directory from the
previous commit(s) where all the jars accumulate.